Some fixes to gameScene and lobby in colliders + % error

+++ Assets/Scripts/GameManager/GameManager.cs
@@ -149,7 +149,7 @@ public class GameManager : MonoBehaviourPunCallbacks, IPunObservable
     void SaveDataToResults()
     {
         //Save data to scriptable object
-        _resultsData.evacuees = _npcManager.peopleEvacuated / _npcManager.GetTotalSpawnedNPCs() * 100;
+        _resultsData.evacuees = (int)((float)_npcManager.peopleEvacuated / (float)_npcManager.GetTotalSpawnedNPCs() * 100.0f);
         _resultsData.domeEnergy = (int)domeEnergy;
         _resultsData.enemiesKilled = enemyKills;
         _resultsData.nDeaths = nDeaths;
